make demo cdc/ct/both aware

+++ sqlserver/03_lakeflow_connect_demo.sh
@@ -52,63 +52,74 @@ export STAGING_SCHEMA=${TARGET_SCHEMA}
 
 # #############################################################################
 
-# create staging and target schemas
-databricks schemas create "$TARGET_SCHEMA" "$TARGET_CATALOG" >/dev/null
-if [[ -n "${DELETE_PIPELINES_AFTER_SLEEP}" ]]; then
-    nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && databricks schemas delete "$TARGET_CATALOG.$TARGET_SCHEMA" >> ~/nohup.out 2>&1 &
+echo -e "\nCreate target and staging schemas \n"
+
+if ! DBX schemas get "$TARGET_CATALOG.$TARGET_SCHEMA"; then
+    if ! DBX schemas create "$TARGET_SCHEMA" "$TARGET_CATALOG"; then
+        cat /tmp/dbx_stderr.$$
+        return 1
+    fi
+    if [[ -n "${DELETE_PIPELINES_AFTER_SLEEP}" ]]; then
+        nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && DBX schemas delete --force "$TARGET_CATALOG.$TARGET_SCHEMA" >> ~/nohup.out 2>&1 &
+    fi
 fi
 
-if [ "$STAGING_CATALOG.$STAGING_SCHEMA" != "$TARGET_CATALOG.$TARGET_SCHEMA" ]; then 
-    databricks schemas create "$STAGING_SCHEMA" "$STAGING_CATALOG" >/dev/null
+if [[ "$TARGET_CATALOG.$TARGET_SCHEMA" != "$STAGING_CATALOG.$STAGING_SCHEMA" ]] && ! DBX schemas get "$STAGING_CATALOG.$STAGING_SCHEMA"; then
+    if ! DBX schemas create "$STAGING_SCHEMA" "$STAGING_CATALOG"; then
+        cat /tmp/dbx_stderr.$$
+        return 1
+    fi
     if [[ -n "${DELETE_PIPELINES_AFTER_SLEEP}" ]]; then
-        nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && databricks schemas delete "$STAGING_CATALOG.$STAGING_SCHEMA" >> ~/nohup.out 2>&1 &
+        nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && DBX schemas delete --force "$STAGING_CATALOG.$STAGING_SCHEMA" >> ~/nohup.out 2>&1 &
     fi
 fi
 
 # #############################################################################
 
-# create
-conn_get_output=$(databricks connections get "$CONNECTION_NAME" 2>/dev/null)
-if [[ -z "$conn_get_output" ]]; then 
-  conn_output=$(databricks connections create --json '{
-    "name": "'"$CONNECTION_NAME"'",
-    "connection_type": "SQLSERVER",
-    "options": {
-      "host": "'"$DB_HOST_FQDN"'",
-      "port": "'"$DB_PORT"'",
-      "trustServerCertificate": "true",
-      "user": "'"$USER_USERNAME"'",
-      "password": "'"${USER_PASSWORD}"'"
-    }
-  }')
+echo -e "\nCreate Connection \n"
+
+# create connection and delete or update
+if ! DBX connections get "$CONNECTION_NAME"; then
+    if ! DBX connections create --json '{
+        "name": "'"$CONNECTION_NAME"'",
+        "connection_type": "SQLSERVER",
+        "options": {
+        "host": "'"$DB_HOST_FQDN"'",
+        "port": "'"$DB_PORT"'",
+        "trustServerCertificate": "true",
+        "user": "'"$USER_USERNAME"'",
+        "password": "'"${USER_PASSWORD}"'"
+        }
+    }'; then
+        cat /tmp/dbx_stderr.$$
+        return 1
+    fi
+    if [[ -n "${DELETE_PIPELINES_AFTER_SLEEP}" ]]; then
+        nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && DBX connections delete "$CONNECTION_NAME" >> ~/nohup.out 2>&1 &
+    fi
 else
   # in case password is updated
-  conn_output=$(databricks connections update "$CONNECTION_NAME" --json '{
-    "options": {
-      "host": "'"$DB_HOST_FQDN"'",
-      "port": "'"$DB_PORT"'",
-      "trustServerCertificate": "true",
-      "user": "'"$USER_USERNAME"'",
-      "password": "'"${USER_PASSWORD}"'"
-    }
-  }')
+  if ! DBX connections update "$CONNECTION_NAME" --json '{
+        "options": {
+        "host": "'"$DB_HOST_FQDN"'",
+        "port": "'"$DB_PORT"'",
+        "trustServerCertificate": "true",
+        "user": "'"$USER_USERNAME"'",
+        "password": "'"${USER_PASSWORD}"'"
+        }
+    }'; then
+        cat /tmp/dbx_stderr.$$
+        return 1
+    fi
 fi
-# needed by the gateway pipeline
-CONNECTION_ID=$(echo "$conn_output" | jq -r '.connection_id')
+CONNECTION_ID=$(jq -r '.connection_id' /tmp/dbx_stdout.$$)
 export CONNECTION_ID
-
-# don't delete connection for now
-if [[ -n $CONNECTION_ID ]]; then
-    echo "CONNECTION_ID=$CONNECTION_ID"
-    echo "Connection ${CONNECTION_NAME}: ${DATABRICKS_HOST}/explore/connections/${CONNECTION_NAME}"
-    echo ""
-else
-    echo "Error: CONNECTION_ID not set"
-fi
 
 # #############################################################################
 
-gw_output=$(databricks pipelines create --json '{
+echo -e "\nCreate Gateway Pipeline \n"
+
+if ! DBX pipelines create --json '{
 "name": "'"$GATEWAY_PIPELINE_NAME"'",
 "gateway_definition": {
   "connection_id": "'"$CONNECTION_ID"'",
@@ -116,29 +127,28 @@ gw_output=$(databricks pipelines create --json '{
   "gateway_storage_schema": "'"$STAGING_SCHEMA"'",
   "gateway_storage_name": "'"$GATEWAY_PIPELINE_NAME"'"
   }
-}')
-GATEWAY_PIPELINE_ID=$(echo "$gw_output" | jq -r '.pipeline_id')
+}'; then
+    cat /tmp/dbx_stderr.$$
+    return 1
+fi
+GATEWAY_PIPELINE_ID="$(jq -r '.pipeline_id' /tmp/dbx_stdout.$$)"
 export GATEWAY_PIPELINE_ID
 
-# print UI URL
-if [[ -n $GATEWAY_PIPELINE_ID ]]; then
-  echo "GATEWAY_PIPELINE_ID=$GATEWAY_PIPELINE_ID"
-    if [[ -n "${STOP_AFTER_SLEEP}" ]]; then 
-        nohup sleep "${STOP_AFTER_SLEEP}" && databricks pipelines stop "$GATEWAY_PIPELINE_ID" >> ~/nohup.out 2>&1 &
-    fi
-    if [[ -n "${DELETE_PIPELINES_AFTER_SLEEP}" ]]; then
-        nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && databricks pipelines delete "$GATEWAY_PIPELINE_ID"  >> ~/nohup.out 2>&1 &
-    fi
-else
-  echo "Error: GATEWAY_PIPELINE_NAME not set"
+if [[ -n "${STOP_AFTER_SLEEP}" ]]; then 
+    nohup sleep "${STOP_AFTER_SLEEP}" && databricks pipelines stop "$GATEWAY_PIPELINE_ID" >> ~/nohup.out 2>&1 &
+fi
+if [[ -n "${DELETE_PIPELINES_AFTER_SLEEP}" ]]; then
+    nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && databricks pipelines delete "$GATEWAY_PIPELINE_ID"  >> ~/nohup.out 2>&1 &
 fi
 
 # #############################################################################
 
+echo -e "\nCreate Ingestion Pipeline \n"
+
 case "${CDC_CT_MODE}" in 
 "BOTH") 
 echo "enabling replication on the schema"
-ig_output=$(databricks pipelines create --json '{
+if ! DBX pipelines create --json '{
 "name": "'"$INGESTION_PIPELINE_NAME"'",
 "continuous": "true",
 "ingestion_definition": {
@@ -152,11 +162,14 @@ ig_output=$(databricks pipelines create --json '{
         }}
     ]
   }
-}')
+}'; then
+    cat /tmp/dbx_stderr.$$
+    return 1
+fi
 ;;
 "CT") 
 echo "enabling replication on the intpk table"
-ig_output=$(databricks pipelines create --json '{
+if ! DBX pipelines create --json '{
 "name": "'"$INGESTION_PIPELINE_NAME"'",
 "continuous": "true",
 "ingestion_definition": {
@@ -171,11 +184,14 @@ ig_output=$(databricks pipelines create --json '{
         }}
     ]
   }
-}')
+}'; then
+    cat /tmp/dbx_stderr.$$
+    return 1
+fi
 ;;
 "CDC") 
 echo "enabling replication on the dtix table"
-ig_output=$(databricks pipelines create --json '{
+if ! DBX pipelines create --json '{
 "name": "'"$INGESTION_PIPELINE_NAME"'",
 "continuous": "true",
 "ingestion_definition": {
@@ -190,7 +206,10 @@ ig_output=$(databricks pipelines create --json '{
         }}
     ]
   }
-}')
+}'; then
+    cat /tmp/dbx_stderr.$$
+    return 1
+fi
 ;;
 *)
 echo "CDC_CT_MODE=${CDC_CT_MODE} must be BOTH|CT|CDC"
@@ -198,55 +217,58 @@ return 1
 ;;
 esac
 
-INGESTION_PIPELINE_ID=$(echo "$ig_output" | jq -r '.pipeline_id')
+INGESTION_PIPELINE_ID=$(jq -r '.pipeline_id' /tmp/dbx_stdout.$$)
 export INGESTION_PIPELINE_ID
 
-ig_get_output=$(databricks pipelines get "$INGESTION_PIPELINE_ID")
-# print UI URL
-if [[ -n $INGESTION_PIPELINE_ID ]]; then
-    echo "INGESTION_PIPELINE_ID=$INGESTION_PIPELINE_ID"
-    if [[ -n "${STOP_AFTER_SLEEP}" ]]; then 
-        nohup sleep "${STOP_AFTER_SLEEP}" && databricks pipelines stop "$INGESTION_PIPELINE_ID" >> ~/nohup.out 2>&1 &
+INGESTION_PIPELINE_CONTINUOUS=$(jq -r '.spec.continuous' /tmp/dbx_stdout.$$)
+export INGESTION_PIPELINE_CONTINUOUS
+
+
+if [[ -n "${STOP_AFTER_SLEEP}" ]]; then 
+    nohup sleep "${STOP_AFTER_SLEEP}" && DBX pipelines stop "$INGESTION_PIPELINE_ID" >> ~/nohup.out 2>&1 &
+fi
+if [[ -n "${DELETE_PIPELINES_AFTER_SLEEP}" ]]; then
+    nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && DBX pipelines delete "$INGESTION_PIPELINE_ID" >> ~/nohup.out 2>&1 &
+fi
+
+# start if not cont
+if [[ "$INGESTION_PIPELINE_CONTINUOUS" == 'false' ]]; then 
+    if ! DBX  pipelines start-update "$INGESTION_PIPELINE_ID"; then
+        cat /tmp/dbx_stderr.$$
+        return 1    
     fi
-    if [[ -n "${DELETE_PIPELINES_AFTER_SLEEP}" ]]; then
-        nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && databricks pipelines delete "$INGESTION_PIPELINE_ID" >> ~/nohup.out 2>&1 &
-    fi
-    if [[ $(echo "$ig_get_output" | jq -r .spec.continuous) == 'false' ]]; then databricks pipelines start-update "$INGESTION_PIPELINE_ID"; fi  
-    echo ""
-else
-    echo "Error: INGESTION_PIPELINE_ID not set"
 fi
 
 # #############################################################################
 
-jobs_output=$(databricks jobs create --json '{
+echo -e "\nCreate Ingestion Pipeline Trigger Jobs  \n"
+
+if ! DBX jobs create --json '{
 "name":"'"$INGESTION_PIPELINE_NAME"'",
 "schedule":{"timezone_id":"UTC", "quartz_cron_expression": "0 5/30 * * * ?"},
 "tasks":[ {
     "task_key":"run_dlt", 
     "pipeline_task":{"pipeline_id":"'"$INGESTION_PIPELINE_ID"'"} } ]
-}')
+}'; then
+    cat /tmp/dbx_stderr.$$
+    return 1    
+fi
 
-INGESTION_JOB_ID=$(echo "$jobs_output" | jq -r '.job_id')
+INGESTION_JOB_ID=$(jq -r '.job_id' /tmp/dbx_stdout.$$)
 export INGESTION_JOB_ID
 
 # print UI URL
-if [[ -n $INGESTION_JOB_ID ]]; then
-    echo "INGESTION_JOB_ID=$INGESTION_JOB_ID"
-    if [[ -n "${STOP_AFTER_SLEEP}" ]]; then 
-        nohup sleep "${STOP_AFTER_SLEEP}" && databricks jobs delete "$INGESTION_JOB_ID" >> ~/nohup.out 2>&1 &
-    fi
-    if [[ -n "${DELETE_PIPELINES_AFTER_SLEEP}" ]]; then
-        nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && databricks jobs delete "$INGESTION_JOB_ID" >> ~/nohup.out 2>&1 &
-    fi
-    if [[ $(echo "$ig_get_output" | jq -r .spec.continuous) == 'false' ]]; then databricks pipelines start-update "$INGESTION_PIPELINE_ID"; fi  
-    echo "Job ${INGESTION_PIPELINE_NAME}: ${DATABRICKS_HOST}/jobs/$INGESTION_JOB_ID"   
-    echo ""
-else
-    echo "Error: INGESTION_JOB_ID not set"
+if [[ -n "${STOP_AFTER_SLEEP}" ]]; then 
+    nohup sleep "${STOP_AFTER_SLEEP}" && DBX jobs delete "$INGESTION_JOB_ID" >> ~/nohup.out 2>&1 &
+fi
+if [[ -n "${DELETE_PIPELINES_AFTER_SLEEP}" ]]; then
+    nohup sleep "${DELETE_PIPELINES_AFTER_SLEEP}" && DBX jobs delete "$INGESTION_JOB_ID" >> ~/nohup.out 2>&1 &
 fi
 
+
 # #############################################################################
+
+echo -e "\nPermission Gateway, Ingestion, Jobs for debug \n"
 
 jobs_pipelines_permission='{
 "access_control_list": [
@@ -266,17 +288,24 @@ databricks permissions update jobs      "$INGESTION_JOB_ID"      --json "$jobs_p
 
 # #############################################################################
 
+echo -e "\Start workload \n"
+
 cat <<EOF | sqlcmd -d "${DB_CATALOG}" -S "${DB_HOST_FQDN},${DB_PORT}" -U "${USER_USERNAME}" -P "${USER_PASSWORD}" -C -l 60 -e >/dev/null 2>&1 &
 while ( 1 = 1 )
 begin
-insert into [${DB_SCHEMA}].[intpk] (dt) values (CURRENT_TIMESTAMP),(CURRENT_TIMESTAMP), (CURRENT_TIMESTAMP)
-insert into [${DB_SCHEMA}].[dtix] (dt) values (CURRENT_TIMESTAMP),(CURRENT_TIMESTAMP),(CURRENT_TIMESTAMP)
-delete from [${DB_SCHEMA}].[intpk] where pk=(select min(pk) from [${DB_SCHEMA}].[intpk])
-update [${DB_SCHEMA}].[intpk] set dt=CURRENT_TIMESTAMP where pk=(select min(pk) from [${DB_SCHEMA}].[intpk])
+IF OBJECT_ID(N'${DB_SCHEMA}.intpk', N'U') IS NOT NULL
+    begin
+    insert into [${DB_SCHEMA}].[intpk] (dt) values (CURRENT_TIMESTAMP),(CURRENT_TIMESTAMP), (CURRENT_TIMESTAMP)
+    delete from [${DB_SCHEMA}].[intpk] where pk=(select min(pk) from [${DB_SCHEMA}].[intpk])
+    update [${DB_SCHEMA}].[intpk] set dt=CURRENT_TIMESTAMP where pk=(select min(pk) from [${DB_SCHEMA}].[intpk])
+    end
+IF OBJECT_ID(N'${DB_SCHEMA}.dtix', N'U') IS NOT NULL
+    insert into [${DB_SCHEMA}].[dtix] (dt) values (CURRENT_TIMESTAMP),(CURRENT_TIMESTAMP),(CURRENT_TIMESTAMP)
 WAITFOR DELAY '00:00:01'
 end
 go
 EOF
+
 LOAD_GENERATOR_PID=$!
 if [[ -n "${STOP_AFTER_SLEEP}" ]]; then 
     nohup sleep "${STOP_AFTER_SLEEP}" && kill -9 "$LOAD_GENERATOR_PID" >> ~/nohup.out 2>&1 &
@@ -288,5 +317,12 @@ echo "Load Generator: started with PID=$LOAD_GENERATOR_PID."
 echo ""
 
 # #############################################################################
+
+echo -e "\n Click on UI \n"
+
+echo -e "Staging schema ${DATABRICKS_HOST}/explore/data/${STAGING_CATALOG}/${STAGING_SCHEMA}"
+echo -e "Target schema ${DATABRICKS_HOST}/explore/data/${TARGET_CATALOG}/${TARGET_SCHEMA}"
+echo -e "Connection ${CONNECTION_NAME}: ${DATABRICKS_HOST}/explore/connections/${CONNECTION_NAME}"
+echo -e "Job ${INGESTION_PIPELINE_NAME}: ${DATABRICKS_HOST}/jobs/$INGESTION_JOB_ID \n"   
 
 databricks pipelines list-pipelines --filter "name like 'robertlee_%'" | jq --arg url "$DATABRICKS_HOST" -r 'sort_by(.name) | .[] | [ .name, .pipeline_id, .state, ($url + "/pipelines/" + .pipeline_id) ] | @tsv'
